Refactor holiday generator into modular package with CLI wrapper
